+++ farben/src/macros/strip.rs
@@ -32,8 +32,11 @@
 /// ```
 #[macro_export]
 macro_rules! ansi_strip {
-    ($($arg:tt)*) => {{
-        $crate::strip_ansi(&format!($($arg)*))
+    ($s:expr) => {{
+        $crate::strip_ansi(&$s.to_string())
+    }};
+    ($fmt:literal $(, $arg:expr)*) => {{
+        $crate::strip_ansi(&format!($fmt $(, $arg)*))
     }};
 }
 
@@ -60,6 +63,9 @@ macro_rules! ansi_strip {
 /// ```
 #[macro_export]
 macro_rules! markup_strip {
+    ($s:expr) => {{
+        $crate::strip_markup(&$s.to_string())
+    }};
     ($($arg:tt)*) => {{
         $crate::strip_markup(&format!($($arg)*))
     }};