+++ testfarbenyay/src/main.rs
@@ -3,6 +3,6 @@ farben::load_styles!();
 
 fn main() {
     init_styles();
-    cprintln!("]]] Bold red");
+    cprintln!("]]]]]] Bold red");
     // cprintln!("\\[ wouldn't work no more");
 }